pyproject.toml
[project]
name = "msv2-inference"
version = "0.1.0"
description = "OpenL3 audio embedding inference service for Knative"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.100.0",
    "librosa>=0.10.0",
    "minio>=7.1.0",
    "numpy<2.0.0",
    "onnxruntime>=1.15.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.6",
    "soundfile>=0.12.0",
    "uvicorn>=0.20.0",
]
[tool.setuptools]
packages = ["api"]
